chore: contextualize Web3ModalProvider
